+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 * Add PayPalHere details
 * Add `networkResponseCode` and `networkResponseText` to transactions and verifications
+* Add `cavv`, `xid`, `ds_transaction_id`, `eci_flag`, and `three_d_secure_version`, to `three_d_secure_info`
+* Add `three_d_secure_info` to credit_card_verification
 
 ## 3.55.0
 * Add `captureId` field to local_payment_details
@@ -11,7 +13,6 @@
 * Add `refundFromTransactionFeeAmount` field to local_payment_details
 * Add `refundFromTransactionFeeCurrencyIsoCode` field to local_payment_details
 * Add `ds_transaction_id` and `three_d_secure_version` to 3DS pass thru fields
-* Add `cavv`, `xid`, `ds_transaction_id`, `eci_flag`, and `three_d_secure_version`, to `three_d_secure_info`
 * Add `payer_info` field to payment_method_nonce details
 * Add more specific timeout errors: (#105 thanks @bhargavrpatel)
 * Add `braintree.exceptions.http.timeout_error.ConnectTimeoutError` (child class of TimeoutError)